# be/docs/swagger.yaml
basePath: /api
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ErrorData'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  models.ExportResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      download_url:
        type: string
      file_size:
        type: integer
      job_id:
        type: string
      record_count:
        type: integer
      status:
        description: '"pending", "processing", "completed", "failed"'
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@form.com
    name: API Support
  description: API for form application with JWT authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: form API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access and refresh tokens
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate tokens (client-side mainly)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get the current authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access and refresh tokens using a valid refresh token
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email, password, and name
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Register a new user
      tags:
      - auth
  /exports/{job_id}/download:
    get:
      consumes:
      - application/json
      description: Download the exported file when the job is completed
      parameters:
      - description: Export job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Download exported file
      tags:
      - export
  /exports/{job_id}/status:
    get:
      consumes:
      - application/json
      description: Check the status of an export job and get download URL when ready
      parameters:
      - description: Export job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ExportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get export job status
      tags:
      - export
  /forms:
    get:
      consumes:
      - application/json
      description: Get all forms created by current user (paginated)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      - default: modified
        description: Sort by (created or modified)
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order (asc or desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all user forms
      tags:
      - forms
    post:
      consumes:
      - application/json
      description: Create a new form (survey or quiz)
      parameters:
      - description: Create form request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateFormRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new form
      tags:
      - forms
  /forms/{form_id}:
    delete:
      consumes:
      - application/json
      description: Delete form (soft delete)
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete form
      tags:
      - forms
    get:
      consumes:
      - application/json
      description: Get form details with all sections and fields
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get form details
      tags:
      - forms
    put:
      consumes:
      - application/json
      description: Update form settings (partial update)
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Update form request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateFormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Update form settings
      tags:
      - forms
  /forms/{form_id}/analytics:
    get:
      consumes:
      - application/json
      description: Retrieve detailed analytics including summary statistics, score
        distribution, time analytics, and question performance
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Start date for analytics period (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: End date for analytics period (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: Group responses by time period
        enum:
        - day
        - week
        - month
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get comprehensive analytics for a form
      tags:
      - analytics
  /forms/{form_id}/analytics/questions:
    get:
      consumes:
      - application/json
      description: Retrieve detailed analytics for each question including answer
        distribution and performance metrics
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get question-level analytics for a form
      tags:
      - analytics
  /forms/{form_id}/analytics/sections:
    get:
      consumes:
      - application/json
      description: Retrieve analytics for each section including average scores and
        completion rates
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get section-level analytics for a form
      tags:
      - analytics
  /forms/{form_id}/analytics/trends:
    get:
      consumes:
      - application/json
      description: Retrieve trend data showing responses over time with grouping options
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Group responses by time period
        enum:
        - day
        - week
        - month
        in: query
        name: group_by
        required: true
        type: string
      - description: Start date for trend analysis (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: End date for trend analysis (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get response trends over time for a form
      tags:
      - analytics
  /forms/{form_id}/duplicate:
    post:
      consumes:
      - application/json
      description: Create a copy of form with all settings and questions
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Duplicate form request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DuplicateFormRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Duplicate form
      tags:
      - forms
  /forms/{form_id}/export:
    get:
      consumes:
      - application/json
      description: Export form responses as CSV, Excel, or JSON with customizable
        columns and data inclusion
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Export format
        enum:
        - csv
        - excel
        - json
        in: query
        name: format
        required: true
        type: string
      - collectionFormat: csv
        description: Specific columns to include
        in: query
        items:
          type: string
        name: columns
        type: array
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: from_date
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: to_date
        type: string
      - default: true
        description: Include detailed answers
        in: query
        name: include_answers
        type: boolean
      - default: false
        description: Include analytics summary
        in: query
        name: include_analytics
        type: boolean
      - description: Filter by pass/fail status
        in: query
        name: is_passed
        type: boolean
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Export form responses in various formats
      tags:
      - export
  /forms/{form_id}/fields:
    get:
      consumes:
      - application/json
      description: Get all fields in form (ordered by order_global)
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: List all fields
      tags:
      - fields
    post:
      consumes:
      - application/json
      description: Add question or content to form
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Create field request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateFieldRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new field
      tags:
      - fields
  /forms/{form_id}/fields/{field_id}:
    delete:
      consumes:
      - application/json
      description: Delete field (associated answers also deleted)
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Field ID
        in: path
        name: field_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete field
      tags:
      - fields
    get:
      consumes:
      - application/json
      description: Get single field details
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Field ID
        in: path
        name: field_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get field details
      tags:
      - fields
    put:
      consumes:
      - application/json
      description: Update field/question details
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Field ID
        in: path
        name: field_id
        required: true
        type: string
      - description: Update field request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateFieldRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Update field
      tags:
      - fields
  /forms/{form_id}/fields/reorder:
    patch:
      consumes:
      - application/json
      description: Reorder fields (change order_global and section assignments)
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Reorder fields request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ReorderFieldsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Reorder fields
      tags:
      - fields
...
  /whitelist/validate/{access_token}:
    get:
      description: Check if token is valid (used by quiz taker) - No authentication
        required
      parameters:
      - description: Access Token
        in: path
        name: access_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Validate access token
      tags:
      - whitelist
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"